Move investment CollectionLists into modules directory
