+++ src/client/modules/Investments/Opportunities/metadata.js
