+++ src/client/modules/Investments/Profiles/transformers.js
@@ -1,7 +1,7 @@
 /* eslint-disable camelcase */
 import { companies } from '../../../../lib/urls'
 
-const { format } = require('../../../../client/utils/date')
+const { format } = require('../../../utils/date')
 
 const transformLargeCapitalProfiles = ({ investor_company, created_on }) => ({
   headingText: investor_company.name,
